requirement.txt
panda
streamlit
numpy
plotly